// src/generated/orval/models/userProfileRequest.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * 모든 API
 * OpenAPI spec version: v0.4
 */
import type { UserProfileRequestPosition } from './userProfileRequestPosition';

export interface UserProfileRequest {
  /** 이메일 */
  email?: string;
  /** 학번 */
  studentNumber: string;
  /** 감자 동아리 기수 */
  gamjaBatch: number;
  /** 포지션/직책 */
  position: UserProfileRequestPosition;
}